Fix error checking in Renderer Init
The code returns true in the check wether the Renderer init was successfull. This not only indicates that the initialization was successful (which it would not be in this case), it also skips the further checks on Mixer and TTF init.
Changing the return value to false fixes this issue.